refactor: 불필요한 인터페이스 분리 삭제
