+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -8,8 +8,8 @@ import nextstep.favorite.infrastructure.FavoriteRepository;
 import nextstep.favorite.presentation.FavoriteRequest;
 import nextstep.member.domain.Member;
 import nextstep.member.infrastructure.MemberRepository;
-import nextstep.subway.domain.PathFinderService;
 import nextstep.subway.domain.Path;
+import nextstep.subway.domain.PathFinderService;
 import nextstep.subway.domain.PathType;
 import nextstep.subway.domain.Station;
 import nextstep.subway.infrastructure.StationRepository;
